Extract TypeDoc from DocExplorer
